adding nav bar
